requirements.txt
# --- Framework principal ---
fastapi

# --- Servidor ---
uvicorn[standard]

# --- Base de datos ---
SQLAlchemy
psycopg2-binary

# --- Migraciones ---
alembic

# --- Validaciones ---
pydantic
pydantic-settings

# --- Seguridad y JWT ---
python-jose
passlib[bcrypt]

# --- Utilidades ---
python-dotenv

# --- Tipos y mejoras ---
typing-extensions

# --- Para CORS ---
starlette
